# src/main/resources/static/swagger-apis/api1/openapi.yaml
openapi: 3.0.0
info:
  title: Maveric Digital Interview Process APIs
  description: REST APIs supporting Maveric Digital Interview process. Mainly contains APIs for candidate registration, login, taking tests, storing test results, viewing candidate details and test reports
  version: 0.0.1
servers:
  - url: http:/localhost:8080/
    description: Local development server
paths:
...
  /candidates:
    get:
      summary: To get the list of all candidates.
      description: API to get the list of all candidates.
      operationId: getCandidates
      tags:
        - Candidates
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: List of all candidates
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    post:
      summary: To register a new candidate.
      description: API to register a new candidate.
      operationId: registerCandidate
      tags:
        - Candidates
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      requestBody:
        description: Candidate details to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '201':    # status code
          description: Candidate successfully registered
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
  /candidates/{candidateId}:
    get:
      summary: To get the details of a candidate
      description: API to get the details of a candidate
      tags:
        - Candidates
      operationId: getCandidateDetails
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: candidateId
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Candidate successfully registered
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    put:
      summary: To update a candidate details
      description: API to update a candidate details
      tags:
        - Candidates
      operationId: updateCandidate
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: candidateId
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      requestBody:
        description: Candidate details to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '200':    # status code
          description: Candidate successfully updated
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: To delete a candidate
      description: API to delete a candidate details
      tags:
        - Candidates
      operationId: deleteCandidate
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: candidateId
          in: path
          description: Candidate ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Candidate successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
  /employees:
    get:
      summary: To get the list of all employees.
      description: API to get the list of all employees.
      operationId: getEmployees
      tags:
        - Employees
      responses:
        '200':
          description: List of all employees
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    post:
      summary: To register a new employee.
      description: API to register a new employee.
      operationId: registerEmployee
      tags:
        - Employees
      requestBody:
        description: Employee details to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':    # status code
          description: Employee successfully registered
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
  /employees/{employeeId}:
    get:
      summary: To get the details of a employee
      description: API to get the details of a employee
      tags:
        - Employees
      operationId: getEmployeeDetails
      parameters:
        - name: employeeId
          in: path
          description: Employee ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Employee successfully registered
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    put:
      summary: To update a employee details
      description: API to update a employee details
      tags:
        - Employees
      operationId: updateEmployee
      parameters:
        - name: employeeId
          in: path
          description: Employee ID
          required: true
          schema:
            type: string
      requestBody:
        description: Employee details to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':    # status code
          description: Employee details successfully updated
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: To delete a employee
      description: API to delete a employee details
      tags:
        - Employees
      operationId: deleteEmployee
      parameters:
        - name: employeeId
          in: path
          description: Employee ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Employee successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    Candidate:
      type: object
      properties:
        _id:
          type: string
          description: The candidate ID.
        firstName:
          type: string
          description: Candidate's first name
        lastName:
          type: string
          description: Candidate's last name
        middleName:
          type: string
          description: Candidate's middle name
        dob:
          type: string
          format: date
          description: Candidate's date of birth
        experience:
          type: number
          description: Candidate's total experience
        skillset:
          type: string
          description: Brief description about Candidate's technical skillset
        panNo:
          type: string
          description: Candidate's PAN number
        email:
          type: string
          description: Candidate's email id
        phoneNo:
          type: number
          description: Candidate's phone number
        noticePeriod:
          type: number
          description: Candidate's notice period in months
        gender:
          type: string
          description: Candidate's gender (Male/Female)
...
    Employee:
      type: object
      properties:
        _id: 
          type: string
          description: Unique ID to identify the employee
        firstName:
          type: string
          description: Employee's first name
        lastName:
          type: string
          description: Employee's last name
        middleName:
          type: string
          description: Employee's middle name
        empId:
          type: string
          description: Employee ID
        password:
          type: string
          description: Hashed password for login
        email:
          type: string
          description: Employee's email id
        phoneNo: 
          type: number
          description: Employee's phone number
        designation:
          type: string
          description: Employee's designation with in the organisation
...